Move configuration loading to its own utility file
